Making segment metadata use long instaed of int since beyond 2 gigs, offset values will reset for int offsets

+++ src/main/java/com/mydb/mydb/entity/SegmentMetadata.java
@@ -15,6 +15,6 @@ public class SegmentMetadata implements Serializable {
   @Serial
   private static final long serialVersionUID = 5388380270261334689L;
 
-  private int offset;
-  private int size;
+  private long offset;
+  private long size;
 }
